streamline testing
This refactors the tests into an easier to read structure.
Arguably, a proc macro could be used to derive many
of the tests `#[variants(i32, u32, i64, u64)]` but I think that may
beyond the scope of this ticket.

I also refactored the way strings are converted, which
also doubly verifies the bounds checking for double conversion.