correctly calculate index length
We had a small bug in our code in the way we calculated the index for
cache insertion and retrieval.

When the `num_chars` was equal to the message length, we were returning
the message length as the index, but in fact still would have 1 element
in the string to return.