SQL-1300: make regex pin to start and end of string
I updated our regex to pin the user pattern to the start and end of the string.
This way, "downtimes" would not match "downtimestatus" , and "status"
would not match "downtimestatus".

More work will need to be done to deconflict between SQL wildcard chars
and allowable names in MongoDB. I've added a comment with the new ticket #