SQL-2051: clean up statements in disconnect if necessary (#208)
* SQL-2051: clean up statements in disconnect if necessary

SQLDisconnect now closes cursors on any open statements as part of its internal logic. This is safe to do since the cursor is an option, and closing it sets the option to None, which will allow Rust to clean up its resources automatically. Iterating over the statements is as safe as possible with no unwraps. The nesting gets a bit deep but it's not too unwieldy.

Additionally, as part of this PR I made a few changes to the common test infra to make it easier to get the default connection string.

* improve tests

* test comment

* update other tests

* pr feedback

* more cleanup

* feedback