SQL-2428: update to proper paths (#261)
* SQL-2428: update to proper paths

This updates the paths in the release to be prepended with "eap".

Additionally, the release version and libv version are fetched slightly differently now.

Most of the changes to the evergreen file are auto formatting, with extraneous spaces removed.

It apears I targeted the docs at the wrong base (master), so they all appear here. The only additions from the approved PR are "Escape Fields" under ## Usage, and ##### Prerequisites under #### Windows Installation

* feedback

* remove download blurb